feat: transfer singUp redirect responsibility to auth pages
