+++ src/app/(auth)/login/page.tsx
@@ -1,48 +1,18 @@
-'use client';
-import React, {Suspense, useContext} from 'react'
+import React, {Suspense} from 'react'
+import Login from "@/app/(auth)/login/Login";
+import {cookies} from "next/headers";
+import {redirect} from "next/navigation";
 
-import {AuthContext, SignInCredentials} from '@/contexts/AuthContext'
-import {SubmitHandler, useForm} from "react-hook-form";
-import Button from "@/components/Button";
-import Link from "next/link";
-import {useSearchParams} from "next/navigation";
-
-const Login = () => {
-  const {signIn} = useContext(AuthContext);
-  const redirect = useSearchParams().get('redirect');
-
-  const {
-    register,
-    handleSubmit,
-  } = useForm<SignInCredentials>()
-
-  const onSubmit: SubmitHandler<SignInCredentials> = async (data: SignInCredentials) => {
-    await signIn(data);
-  }
+// had to wrap my login and register client components into a side component, because the client wouldn't redirect first try to another server component
+const Page = async ({searchParams}: { searchParams: { [key: string]: string | undefined } }) => {
+  if (cookies().get("ovb.token"))
+    redirect(searchParams.redirect || "/");
 
   return (
-    <div
-      className="container flex justify-center items-center w-[300px] lg:w-[400px] xl:w-[500px] min-h-[calc(100vh-230px)]">
-      <div className="m-2 w-full">
-        <h1 className="text-xl mb-8 font-bold uppercase">LOGIN</h1>
-        <form onSubmit={handleSubmit(onSubmit)} className="text-lg">
-          <input type="text" {...register('email')} placeholder="Email"
-                 className="block mb-2 border-b border-b-gray-200 w-full"/>
-          <input type="password" {...register('password')} placeholder="Password"
-                 className="block mb-4 border-b border-b-gray-200 w-full"/>
-          <Button primary={true}>
-            Login
-          </Button>
-        </form>
-        <Suspense>
-          <div className="mt-2 text-gray-500 text-sm">
-            Not registered yet? <Link href={`/register${redirect ? '?redirect=' + encodeURIComponent(redirect) : ''}`}
-                                      className="text-blue-600">Sign up</Link>
-          </div>
-        </Suspense>
-      </div>
-    </div>
+    <Suspense fallback={<div>loading...</div>}>
+      <Login/>
+    </Suspense>
   )
 }
 
-export default Login;
+export default Page;
